Create WindowsCLibrary
