+++ src/main/java/org/fusesource/jansi/ffm/AnsiConsoleSupportFfm.java
@@ -18,16 +18,11 @@ package org.fusesource.jansi.ffm;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.lang.foreign.Arena;
-import java.lang.foreign.FunctionDescriptor;
-import java.lang.foreign.GroupLayout;
-import java.lang.foreign.Linker;
-import java.lang.foreign.MemoryLayout;
 import java.lang.foreign.MemorySegment;
 import java.lang.foreign.ValueLayout;
-import java.lang.invoke.MethodHandle;
-import java.lang.invoke.VarHandle;
 
 import org.fusesource.jansi.AnsiConsoleSupport;
+import org.fusesource.jansi.internal.OSInfo;
 import org.fusesource.jansi.io.AnsiProcessor;
 
 import static org.fusesource.jansi.ffm.Kernel32.*;
@@ -40,68 +35,11 @@ public class AnsiConsoleSupportFfm implements AnsiConsoleSupport {
 
     @Override
     public CLibrary getCLibrary() {
-        return new CLibrary() {
-            private static final int TIOCGWINSZ;
-            private static final GroupLayout wsLayout;
-            private static final MethodHandle ioctl;
-            private static final VarHandle ws_col;
-            private static final MethodHandle isatty;
-
-            static {
-                String osName = System.getProperty("os.name");
-                if (osName.startsWith("Linux")) {
-                    String arch = System.getProperty("os.arch");
-                    boolean isMipsPpcOrSparc =
-                            arch.startsWith("mips") || arch.startsWith("ppc") || arch.startsWith("sparc");
-                    TIOCGWINSZ = isMipsPpcOrSparc ? 0x40087468 : 0x00005413;
-                } else if (osName.startsWith("Solaris") || osName.startsWith("SunOS")) {
-                    int _TIOC = ('T' << 8);
-                    TIOCGWINSZ = (_TIOC | 104);
-                } else if (osName.startsWith("Mac") || osName.startsWith("Darwin")) {
-                    TIOCGWINSZ = 0x40087468;
-                } else if (osName.startsWith("FreeBSD")) {
-                    TIOCGWINSZ = 0x40087468;
-                } else {
-                    throw new UnsupportedOperationException();
-                }
-
-                wsLayout = MemoryLayout.structLayout(
-                        ValueLayout.JAVA_SHORT.withName("ws_row"),
-                        ValueLayout.JAVA_SHORT.withName("ws_col"),
-                        ValueLayout.JAVA_SHORT,
-                        ValueLayout.JAVA_SHORT);
-                ws_col = wsLayout.varHandle(MemoryLayout.PathElement.groupElement("ws_col"));
-                Linker linker = Linker.nativeLinker();
-                ioctl = linker.downcallHandle(
-                        linker.defaultLookup().find("ioctl").get(),
-                        FunctionDescriptor.of(
-                                ValueLayout.JAVA_INT, ValueLayout.JAVA_INT, ValueLayout.JAVA_LONG, ValueLayout.ADDRESS),
-                        Linker.Option.firstVariadicArg(2));
-                isatty = linker.downcallHandle(
-                        linker.defaultLookup().find("isatty").get(),
-                        FunctionDescriptor.of(ValueLayout.JAVA_INT, ValueLayout.JAVA_INT));
-            }
-
-            @Override
-            public short getTerminalWidth(int fd) {
-                try (Arena session = Arena.ofConfined()) {
-                    MemorySegment segment = session.allocate(wsLayout);
-                    int res = (int) ioctl.invoke(fd, (long) TIOCGWINSZ, segment);
-                    return (short) ws_col.get(segment);
-                } catch (Throwable e) {
-                    throw new RuntimeException("Unable to ioctl(TIOCGWINSZ)", e);
-                }
-            }
-
-            @Override
-            public int isTty(int fd) {
-                try {
-                    return (int) isatty.invoke(fd);
-                } catch (Throwable e) {
-                    throw new RuntimeException("Unable to call isatty", e);
-                }
-            }
-        };
+        if (OSInfo.isWindows()) {
+            return new WindowsCLibrary();
+        } else {
+            return new PosixCLibrary();
+        }
     }
 
     @Override
@@ -157,4 +95,5 @@ public class AnsiConsoleSupportFfm implements AnsiConsoleSupport {
             }
         };
     }
+
 }
